pyproject.toml
[project]
name = "foundrysdkexplorations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "azure-ai-projects==2.0.0b4",
    "python-dotenv>=1.2.2",
]
